requirements.txt
# PolySleuth Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
web3>=6.11.0
requests>=2.31.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
websockets>=12.0
